feat(perf): apply next/dynamic code splitting to admin and inbox pages
Extract CategoryFormModal and TaskDetailPanel into separate chunks via
next/dynamic, reducing initial bundle size for faster Cloud Run cold starts.
