+++ package.json
@@ -16,6 +16,7 @@
     "@supabase/ssr": "^0.10.2",
     "@supabase/supabase-js": "^2.104.1",
     "clsx": "^2.1.1",
+    "lightningcss": "^1.32.0",
     "lucide-react": "^1.7.0",
     "next": "^16.2.4",
     "react": "19.2.4",
